Create main.clj
Function translate manipulates the given message in a way so that there are no illegal simbols and that all spaces are replaced with underscores. Interesting fact with the given task is that even though it's been stated that spaces must be replaced with underscores, including underscores DIRECTLY in the message is illegal. Could be an oversight, who knows, certainly not me.

Function rail retrieves the resulting string fragment in a specific rail, e.g, rail one, rail two, rail three, etc.

Function combine, as the name implies, combines the results of all the rails resulting in a finished, encrypted message.

Function encrypt is just there, so that the user doesn't have to input the two extra parameters in function combine. The parameters are there to aid recursion but are not meant to be "fed" into the first call by the user, so the function encrypt ensures, that the parameters result and row are always "" and 1, respectively.