Update main.clj
Essentially the first version of the decryption function has been created. It has passed only one test so far and it will further be adjusted to be more, well, universal. For now, it works perfectly with strings where the unfinished cycle finishes on the final row, for example, 15 symbols and 3 rails.